Move the OpenStack pkg up a level
We may need to implement two versions of the CSI driver,
one for standalone Cinder and one for a standard Cloud Deployment.

This change just elevages the OpenStack pkg in csi up a level so
that it can be shared.  Eventually, it might be worth considering
doing some work to just share the existing:
  cloud-provider-openstack/pkg/provider implementation

For now however that's significantly different and we'd need to
either add a number of things and modify it a bit, or modify the
the CSI work significantly.  Adding/modifying the higher level
provider may be the best bet in the long run, but for now the
needed pieces for CSI are pretty minimal so let's just keep them
isolated for now and not risk breaking existing deployments.